fixing warnings in previous patch (typecasting pointer to long)

+++ xlators/cluster/afr/src/afr-transaction.c
@@ -887,7 +887,7 @@ int afr_lock_rec (call_frame_t *frame, xlator_t *this, int child_index)
 
 int32_t afr_lock (call_frame_t *frame, xlator_t *this)
 {
-        frame->root->pid = frame->root;
+        frame->root->pid = (long) frame->root;
 
 	return afr_lock_rec (frame, this, 0);
 }
